Update query-generator.js
When we use view in model, and update record, mssql returns error Msg 404, Level 16, State 1 "The column reference "inserted.<COLUMN_NAME>" is not allowed because it refers to a base table that is not being modified in this statement" for each data column in model, because of "... OUTPUT INSERTED.* into ...", that sequelize adds without any possibility to disable it. And there is no sense at all to use " OUTPUT INSERTED.* " in update clause, because we get only count of processed records: const [count] = await this.update(props, opts).